refactor: 수정/삭제 모달 아이콘 컴포넌트로 수정
