+++ src/components/feature/EditDeleteIcon/EditDelete.style.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const RightIcon = styled.div`
+  margin-left: auto;
+  color: var(--color-gray-light);
+  transition: color 0.25s ease;
+  cursor: pointer;
+  position: relative;
+  &:hover {
+    color: var(--color-gray-dark);
+  }
+`;
